Extracts test methods out of deployment code.
In order to make this go smoothly, I greatly increased the strictness of TS, and then fixed all of the errors.  I also upgraded to TS 2.6 which gave us some new static analysis.

This addresses the biggest concern for getting deployment CI working properly, which was that there was test code happening as part of deployment (creating a market).